requirements.txt
azure-functions==1.2.0
pymongo==3.10.1
